Moved SizeUp related logic from Game1 to BrickManager.
SizeUps now correctly flagged for removal if they have applied their effect or if they go outside the play area.

Added Geters for effectApplied bool in the different powerup classes.

Removed test SizeUp and LifeUp from Game1.

Implemented the same logic for LifeUps as for SizeUps in BrickManager and Game1.

Changed the spawn chance for variable for LifeUps to not spawn at the same time as SizeUps.

Began implementing handling of ShitShooters in similar fashion as the powerups.